+++ tests/artifact_storage/basic_test.py
@@ -3,27 +3,37 @@ import tempfile
 from aimrecords import Storage
 
 
-class TestBasicStuff(object):
+class TestMulitpleArtifacts(object):
     def test_simple_int(self):
         len = 1000
 
         with tempfile.TemporaryDirectory() as temp_dir:
-            aim_storage = Storage(temp_dir)
+            storage_writer = Storage(temp_dir, 'w')
 
-            with aim_storage.write('loss') as writer:
-                for i in range(len):
-                    writer.append_record(str(i).encode())
+            storage_writer.open('loss')
+            for i in range(len):
+                storage_writer.append_record('loss', str(i).encode())
+            storage_writer.close('loss')
 
-            with aim_storage.read('loss') as reader:
-                assert reader.records_num == len
-                for i, record in enumerate(reader[:]):
-                    assert i == int(record.decode())
+            storage_writer.open('accuracy')
+            for i in range(len, 2*len):
+                storage_writer.append_record('accuracy', str(i).encode())
+            storage_writer.close('accuracy')
 
-            with aim_storage.write('accuracy') as writer:
-                for i in range(len, 2*len):
-                    writer.append_record(str(i).encode())
+            del storage_writer
 
-            with aim_storage.read('accuracy') as reader:
-                assert reader.records_num == len
-                for i, record in enumerate(reader[:]):
-                    assert i+len == int(record.decode())
+            storage_reader = Storage(temp_dir, 'r')
+
+            storage_reader.open('loss')
+            assert storage_reader.get_length('loss') == len
+            for i, record in enumerate(storage_reader.read_records('loss')):
+                assert i == int(record.decode())
+            storage_reader.close('loss')
+
+            storage_reader.open('accuracy')
+            assert storage_reader.get_length('accuracy') == len
+            for i, record in enumerate(storage_reader.read_records('accuracy')):
+                assert i + len == int(record.decode())
+            storage_reader.close('accuracy')
+
+            del storage_reader