+++ tests/record_storage/basic_test.py
@@ -10,32 +10,32 @@ class TestBasicStuff(object):
         with tempfile.TemporaryDirectory() as temp_dir:
             path = os.path.join(temp_dir, 'loss')
 
-            writer = Writer(path)
-            len = 10000
-            for i in range(len):
-                writer.append_record(str(i).encode())
-            writer.finalize()
+            writer = Writer(path, compression=None)
+            length = 1000
+            for index in range(length):
+                writer.append_record(str(index).encode())
+            writer.close()
 
             reader = Reader(path)
-            assert reader.records_num == len
+            assert reader.records_num == length
 
-            for index in range(len):
+            for index in range(length):
                 assert index == int(reader.get(index).decode())
 
     def test_simple_binary(self):
         with tempfile.TemporaryDirectory() as temp_dir:
             path = os.path.join(temp_dir, 'loss')
 
-            writer = Writer(path)
-            len = 10000
-            for index in range(len):
+            writer = Writer(path, compression=None)
+            length = 5000
+            for index in range(length):
                 entry = (str(index) * index).encode()
                 writer.append_record(entry)
-            writer.finalize()
+            writer.close()
 
             reader = Reader(path)
-            assert reader.records_num == len
+            assert reader.records_num == length
 
-            for index in range(len):
+            for index in range(length):
                 entry = (str(index) * index).encode()
                 assert entry == reader.get(index)